refactor(crux): consolidate types, centralize config, fix README
- Create shared CommandOptions type in crux/lib/command-types.ts and
  update 14 command files to use it instead of duplicated declarations
- Centralize wiki-server timeout constants in crux/lib/config.ts,
  replacing duplicated values in client.ts and sync-common.ts
- Fix crux/README.md: remove duplicate edit-log entry, add 27 missing
  domain references, update page count from ~625 to ~660
- Fix --no-limit flag in issues create (kebab-case wasn't recognized
  after crux.mjs converts to camelCase)

Closes #1754
Closes #1756
Closes #1757
